+++ CHANGELOG.md
@@ -23,6 +23,7 @@ if the selection is initiated by a long press
 ### Fixed
 - Crash when clicking the "Apply" button on the search screen twice
 - Media viewer failing to load very big images
+- Missing player controls when viewing video in rare cases
 
 ## [1.16.0] - 2023-07-31